requirements.txt
websocket-client==1.2.3
python-dotenv==0.19.2
websocket-server==0.6.4
